Fix lint errors

+++ src/consts.rs
@@ -11,8 +11,8 @@ pub const TS_SHIFT: i32 = SEQ_BITS + WID_BITS;
 pub const WID_SHIFT: i32 = SEQ_BITS;
 
 // maximize timestamp value
-pub const MAX_TS: u64 = (-1 as i64 ^ (-1 << TS_BITS)) as u64;
+pub const MAX_TS: u64 = (-1_i64 ^ (-1 << TS_BITS)) as u64;
 // maximize worker id value
-pub const MAX_WID: u16 = (-1 as i16 ^ (-1 << WID_BITS)) as u16;
+pub const MAX_WID: u16 = (-1_i16 ^ (-1 << WID_BITS)) as u16;
 // maximize sequence value
-pub const MAX_SEQ: u16 = (-1 as i16 ^ (-1 << SEQ_BITS)) as u16;
+pub const MAX_SEQ: u16 = (-1_i16 ^ (-1 << SEQ_BITS)) as u16;
